requirements.txt
pydantic==2.5.2
pandas==2.1.3
ipykernel==6.27.1
pyyaml==6.0.1
hydra-core==1.3.2
hydra_colorlog==1.2.0